pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["backwardspy <backwardspy@pigeon.life>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.109.0"
python-slugify = "^8.0.3"


[tool.poetry.group.dev.dependencies]
uvicorn = "^0.27.0.post1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
